% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\docType{data}
\name{survey}
\alias{survey}
\title{Metadata for dats set Responses to ... survey}
\format{
A tibble with one observation for each paper, and the following columns:
\describe{
\item{paper_id}{Unique identifier that can link to full text and citations.
SHA of the paper PDF.}
\item{source}{Source (e.g. pubmed, CZI...)}
\item{title}{Title}
\item{doi}{Digital Object Identifier}
\item{pmcid}{pmcid}
\item{pubmed_id}{PubMed ID}
\item{license}{License}
\item{abstract}{Abstract}
\item{publish_time}{Publication year}
\item{authors}{Authors}
\item{journal}{Journal}
\item{microsoft_academic_paper_id}{Microsoft Academic Paper ID}
\item{who}{CovidenceWHO}
\item{has_full_text}{Does it have full text}
}
}
\usage{
survey
}
\description{
type, site and place in the x-survey dataset. this comes from the CSV file
the unprocessed dataset as exported from Kobo Toolbox
}
\details{
raw: Mzedi_survey_-\emph{all_versions}-\emph{False}-_2022-11-28-08-50-18.csv
}
\examples{

library(dplyr)

# What are the most common journals?
survey \%>\% 
   select(1:3)

}
\seealso{
\url{ghe.ethz.ch}
}
\keyword{datasets}
